third_party/blink/web_tests/external/wpt/css/css-display/accessibility/display-contents-role-and-label-expected.txt

This is a testharness.js-based test.
Found 3 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] td as child of tr with display: contents, within table with display: flex, has cell role
  assert_equals: <td data-expectedrole="cell" data-testname="td as child of tr with display: contents, within table with display: flex, has cell role" class="ex-role">x</td> expected "cell" but got "gridcell"
[FAIL] td as child of tr with display: contents, within table with role=table with display: flex, has cell role
  assert_equals: <td data-expectedrole="cell" data-testname="td as child of tr with display: contents, within table with role=table with display: flex, has cell role" class="ex-role">x</td> expected "cell" but got "gridcell"
[FAIL] td within tr in table with role table, all with display: contents, has cell role
  assert_equals: <td style="display: contents;" data-expectedrole="cell" data-testname="td within tr in table with role table, all with display: contents, has cell role" class="ex-role">x</td> expected "cell" but got "gridcell"
Harness: the test ran to completion.
